maps, bugfixes, dyn. music load, safari fixes

+++ srcWebClient/libs/lawnchair_indexedDB.js
@@ -119,7 +119,8 @@ Lawnchair.adapter('indexed-db', (function(){
         }
       }
       request.onerror = function(ev) {
-        if (request.errorCode === getIDBDatabaseException().VERSION_ERR) {
+        if ( getIDBDatabaseException() && 
+          request.errorCode === getIDBDatabaseException().VERSION_ERR) {
           // xxx blow it away
           self.idb.deleteDatabase(self.name);
           // try it again.
